[MOV] survey: move question answer in own file
task-3339514
