+++ addons/survey/models/__init__.py
@@ -7,6 +7,7 @@ from . import ir_http
 from . import res_lang
 from . import res_partner
 from . import survey_question
+from . import survey_question_answer
 from . import survey_survey
 from . import survey_survey_template
 from . import survey_user_input
